requirements.txt
openai>=1.12.0
pydantic-ai
google-generativeai>=0.3.2
python-dotenv>=1.0.0
pydantic>=2.6.1
pydantic-graph>=0.1.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
httpx>=0.26.0
typing-extensions>=4.9.0
pathlib>=1.0.1
rich>=13.7.0
tqdm>=4.66.1